Fix a weird race condition.
I still don't understand why, but sometimes, the old code would have an
incorrect value for the latitude on the GPU. Letting the GPU-kernel-calling C++
function pass its own float to the kernel seems to fix the problem.

+++ examples/analytic_gpu.c
@@ -88,7 +88,7 @@ int main(int argc, char *argv[]) {
     // Number of specified amps per tile.
     int num_amps = 16;
     // MWA latitude
-    double latitude_rad = -0.4660608448386394;
+    FLOAT latitude_rad = -0.4660608448386394;
     // Should we normalise the beam response?
     int norm_to_zenith = 1;
